UIMA-48: fixed some trivial compiler warnings
git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/uimaj/trunk/uimaj-core@478180 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/uima/cas/impl/LongHeap.java
@@ -51,7 +51,7 @@ final class LongHeap extends CommonAuxHeap {
   }
 
   void resetToZeros() {
-    Arrays.fill(this.heap, 0, this.heapPos, (long) NULL);
+    Arrays.fill(this.heap, 0, this.heapPos, NULL);
   }
 
   // Getters
